requirements.txt
djitellopy
networkx
python-box
torch
transformers
torchaudio
sentencepiece
datasets
pyaudio
sounddevice
pyphonetics